+++ scripts/OUTRIDER/outrider.R
@@ -28,6 +28,7 @@ register(MulticoreParam(params$cpu, params$cpu * 2, progressbar = TRUE))
 # -----------------------------------------------------------------------------
 candidates = read.csv(params$candidate_genes)
 candidates_extra = read.table(params$candidate_genes_extra,comment.char = "#",header = T ,sep = ',');candidates_extra[is.na(candidates_extra)] = ''
+candidates_extra = candidates_extra[, colnames(candidates)]
 candidates = rbind(candidates, candidates_extra)
 candidates$ensembl_proband = apply(
   candidates[, colnames(candidates) %in% c('ensembl', 'proband')],