+++ scripts/Shiny/global.R
@@ -77,6 +77,7 @@ transcripts_named_filtered_ggplot = read.csv(file.path(params$datadir, 'transcri
 fc_exons_tpm_ggplot = read.csv(file.path(params$datadir, 'fc_exons_tpm_ggplot.tsv'), sep = '\t', check.names = FALSE)
 candidates = read.csv(file.path(params$datadir, 'input/candidate_genes.csv'))
 candidates_extra = read.table(file.path(params$datadir, 'input/candidate_genes_extra.csv'),comment.char = "#",header = T ,sep = ',');candidates_extra[is.na(candidates_extra)] = ''
+candidates_extra = candidates_extra[, colnames(candidates)]
 candidates = rbind(candidates, candidates_extra)
 candidates = candidates[!grepl('bc',candidates$proband),]