+++ scripts/consensus/consensus.R
@@ -19,6 +19,7 @@ params$candidate_genes_extra = args[7]
 
 candidates = read.csv(params$candidates)
 candidates_extra = read.table(params$candidate_genes_extra,comment.char = "#",header = T ,sep = ',');candidates_extra[is.na(candidates_extra)] = ''
+candidates_extra = candidates_extra[, colnames(candidates)]
 candidates = rbind(candidates, candidates_extra)
 
 # Identify the MANE transcript for the candidate gene